pyproject.toml
[project]
name = "tau2-green-agent"
version = "0.1.0"
description = "Tau2 benchmark green agent for AgentBeats"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "litellm>=1.65.0",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "tau2 @ git+https://github.com/sierra-research/tau2-bench.git@337326e62d8e0ca74c353b004a9c5d748e0ba914",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
